code/bots/requirements.txt
# Enterprise Telegram Bots - Requirements
# Version: 9.0.0
# Python: 3.8+

# Core
python-telegram-bot==20.7
asyncio==3.4.3
aiohttp==3.9.1

# System monitoring
psutil==5.9.6

# HTTP requests
requests==2.31.0

# Configuration
python-dotenv==1.0.0
pyyaml==6.0.1

# Database
jsonschema==4.20.0

# Utilities
python-dateutil==2.8.2

# Optional: Advanced features
# prometheus-client==0.19.0  # Для метрик
# sentry-sdk==1.39.1         # Для error tracking
# redis==5.0.1               # Для кеширования
